deep_researcher/pyproject.toml
[project]
name = "deep-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.1",
    "smolagents[litellm,mcp,openai]>=1.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
